Add pill classes
The pill colors work the same way the button colors do, but the sizes work differently. The approach in the pill classes is better, because it defines the height, padding and font-size based on one size. This is therefore much more precise. 

The line-height attribute has also been set in the pill classes.