Implement round button classes
Removed the pill classes and combined the pill classes with the button classes. The pill class can now be accessed via an option for the button class: 'is-pill'. The same principle has been applied for the round button class. All the button sizes (default, pill and round) are defined in the button settings file. The icon classes have also been adapted. The icon size is now more distributed and the icons are now displayed with block style. This allows centering the icons in the button. The drawback is, that the icon needs a wrapper class in order to stay inline with text. But this will not happen often, because if you use the icon in a button context (inside a button), then the button is the wrapper.