Add disabled state to toggle switch classes
The disabled state here is not a selector but a class. This needs to be done like this, because the native element of the toggle switch is a div, which does not accept the disabled attribute. The new class sets the cursor to not-allowed. The logic in the component makes sure, that the switch does not react on clicks/toggles.